# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lettura API",
    "description": "Lettura is a wallabag-inspired read-it-later service. Endpoints below are stable; unannotated routes exist but their schema is not yet machine-readable.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/entries/bulk/archive": {
      "post": {
        "tags": [
          "bulk"
        ],
        "operationId": "bulk_archive",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk archive result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/entries/bulk/archive-by-ids": {
      "post": {
        "tags": [
          "bulk"
        ],
        "operationId": "bulk_archive_by_ids",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkArchiveByIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk archive by IDs result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/entries/bulk/delete-by-ids": {
      "post": {
        "tags": [
          "bulk"
        ],
        "operationId": "bulk_delete_by_ids",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteByIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk delete by IDs result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/entries/bulk/star": {
      "post": {
        "tags": [
          "bulk"
        ],
        "operationId": "bulk_star",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk star result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/entries/bulk/tag": {
      "post": {
        "tags": [
          "bulk"
        ],
        "operationId": "bulk_tag_add",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk tag result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/entries/bulk/tag-by-ids": {
      "post": {
        "tags": [
          "bulk"
        ],
        "operationId": "bulk_tag_by_ids",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkTagByIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk tag by IDs result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/entries/bulk/untag": {
      "post": {
        "tags": [
          "bulk"
        ],
        "operationId": "bulk_untag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUntagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk untag result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/entries/bulk/untag-by-ids": {
      "post": {
        "tags": [
          "bulk"
        ],
        "operationId": "bulk_untag_by_ids",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUntagByIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk untag by IDs result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/entries/{entry_id}/tags/{tag_id}": {
      "delete": {
        "tags": [
          "tags"
        ],
        "operationId": "remove_tag_from_entry",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "description": "Entry ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tag_id",
            "in": "path",
            "description": "Tag ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag removed from entry"
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "404": {
            "description": "Entry or tag not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkArchiveByIdsRequest": {
        "type": "object",
        "required": [
          "entry_ids"
        ],
        "properties": {
          "entry_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "BulkDeleteByIdsRequest": {
        "type": "object",
        "required": [
          "entry_ids"
        ],
        "properties": {
          "entry_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "BulkResult": {
        "type": "object",
        "required": [
          "matched",
          "updated",
          "ids"
        ],
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "matched": {
            "type": "integer",
            "minimum": 0
          },
          "updated": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "BulkStateRequest": {
        "type": "object",
        "required": [
          "filter",
          "value"
        ],
        "properties": {
          "dry_run": {
            "type": "boolean"
          },
          "filter": {
            "$ref": "#/components/schemas/ListParams"
          },
          "max": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "value": {
            "type": "boolean"
          }
        }
      },
      "BulkTagByIdsRequest": {
        "type": "object",
        "required": [
          "entry_ids",
          "tags"
        ],
        "properties": {
          "entry_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BulkTagRequest": {
        "type": "object",
        "required": [
          "filter"
        ],
        "properties": {
          "add": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dry_run": {
            "type": "boolean"
          },
          "filter": {
            "$ref": "#/components/schemas/ListParams"
          },
          "max": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "BulkUntagByIdsRequest": {
        "type": "object",
        "required": [
          "entry_ids",
          "tags"
        ],
        "properties": {
          "entry_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BulkUntagRequest": {
        "type": "object",
        "required": [
          "filter"
        ],
        "properties": {
          "dry_run": {
            "type": "boolean"
          },
          "filter": {
            "$ref": "#/components/schemas/ListParams"
          },
          "max": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "remove": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ListParams": {
        "type": "object",
        "properties": {
          "before": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Return entries created strictly before this timestamp."
          },
          "cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque cursor for keyset pagination. When present, page/OFFSET is ignored\nand the result is keyed on `(created_at, id) < cursor` ordering."
          },
          "domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "exclude_tag": {
            "type": [
              "string",
              "null"
            ],
            "description": "Comma-separated tag labels to exclude; entry must NOT have any of these tags."
          },
          "fields": {
            "type": [
              "string",
              "null"
            ],
            "description": "Field-projection hint — placeholder for Task 15, not used in query construction."
          },
          "is_archived": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "is_read": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Alias for `is_archived` (CLI Filter DSL compatibility).\n`is_read=true` → `is_archived = true`, `is_read=false` → `is_archived = false`."
          },
          "is_starred": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "page": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "per_page": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "search": {
            "type": [
              "string",
              "null"
            ],
            "description": "Search query. Handled by tantivy in the API handler; unused in list_entries itself."
          },
          "since": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Return entries created at or after this timestamp."
          },
          "tag": {
            "type": [
              "string",
              "null"
            ],
            "description": "Comma-separated tag labels; AND semantics (entry must have ALL listed tags)."
          },
          "untagged": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "When true, return only entries with no tags."
          }
        }
      },
...
    }
  },
...
}